Add staticprefix and the relevant {{ static }} template function
Add a new configuration key, staticprefix, which contains the prefix for
accessing static contents in staticdir.

The value of staticprefix is available in templates through the function
{{ static "string" }}, where `string` is the name of the resource in
staticdir (for example {{ static "custom.css" }}).

The context automatically configures the http.router to serve the
contents in staticdir if staticdir is not empty. Otherwise the static
content must be served by other means (for example by a webserver).

The {{ static }} function in templates will work in both cases.